moved dashboardEvents function to hideCreate.js file that I added within the events folder.
