+++ src/javascripts/helpers/auth/signOut.js
@@ -1,6 +1,7 @@
 import firebase from 'firebase/app';
 import 'firebase/auth';
-import { dashboardEvents, domEvents } from '../../events/domEvents';
+import domEvents from '../../events/domEvents';
+import dashboardEvents from '../../events/hideCreate';
 
 const signOut = () => {
   document.querySelector('body').removeEventListener('click', domEvents);
